test: rename tests/process_group_helpers.py back to _process_group_helpers.py
Resolve F0: the cross-domain rename dropped the leading underscore that
marks helper modules throughout this repo (``tests/_codex_feature_policy.py``,
``tests/skills/_skill_text_helpers.py``, etc.). Restore the underscore so
the file's helper role is still encoded in its name, and update the
single consumer in ``tests/skills/test_ticket_grouper_live_behavior.py``.
